chore: clean comment card imports
